requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2024.6.2
charset-normalizer==2.0.12
flake8==3.9.2
flake8-docstrings==1.6.0
idna==3.7
mccabe==0.6.1
pycodestyle==2.7.0
pydocstyle==6.3.0
pyflakes==2.3.1
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2024.1
requests==2.26.0
six==1.16.0
snowballstemmer==2.2.0
tornado==6.4
tzlocal==5.2
urllib3==1.26.18
